% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotateHR.R
\name{annotateHR}
\alias{annotateHR}
\title{Annotate data with heart rate}
\usage{
annotateHR(Rall, beat_matrix, data)
}
\arguments{
\item{Rall}{data frame with information about (detected) heart beats with columns \itemize{\item \code{Rtrue_idx}: the indexes of R peaks in \code{signal}, \item \code{Rtrue_sec}: the time moments of \code{Rtrue_idx} in seconds, \item \code{RtoRext}: R-R intervals in number of samples (with the starting zero), \item \code{RtoR_secext}: R-R intervals in seconds (with the starting zero).},}

\item{beat_matrix}{output of the \code{\link{HRdistribution}} function containing distribution of heart rate values,}

\item{data}{data frame containing ECG, EEG and GPS measurements}
}
\value{
\code{data}: original data frame complemented with heart rate color code (for plotting).
}
\description{
This function adds factorized code to data points according to heart rate.
}
\examples{
load(system.file("extdata", "ecg.Rda", package="heartBeat"))
load(system.file("extdata", "Rall.Rda", package="heartBeat"))
load(system.file("extdata", "beat_matrix.Rda", package="heartBeat"))
ecg <- annotateHR(Rall, beat_matrix, ecg)

load(system.file("extdata", "ecg.Rda", package="heartBeat"))
result <- heart_beat(ecg)
beat_matrix <- HRdistribution(result$Rall, data$ecg, 75, 30, training = FALSE)
ecg <- annotateHR(Rall, beat_matrix, ecg)
}
\keyword{annotate}
\keyword{heart}
\keyword{rate,}

