% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heart_beat.R
\name{heart_beat}
\alias{heart_beat}
\title{Heart beat detector}
\usage{
heart_beat(data, SampleFreq = 250, thr = 9)
}
\arguments{
\item{data}{Data frame with ECG data.}

\item{SampleFreq}{The sampling frequency in Hz of ECG data. The default setting is 250 Hz.}

\item{thr}{Fixed threshold value used to detect heart beats from wavelet coefficients. The default value is 9.}
}
\value{
The function returns a data object called "result" with 4 elements:
 \itemize{
  \item \code{signal}: the ECG data frame that was analyzed,
  \item \code{coeff}: the coefficients of the 2nd level wavelet decomposition of the ECG signal (\code{signal$ecg}),
  \item \code{R}: the indexes of R peaks in \code{coeff},
  \item \code{Rall}: the data frame with the heart-beat detection results with columns:
   \itemize{
   \item \code{Rtrue_idx}: the indexes of R peaks in \code{signal},
   \item \code{Rtrue_sec}: the time moments of \code{Rtrue_idx} in seconds,
   \item \code{RtoRext}: R-R intervals in number of samples (with the starting zero),
   \item \code{RtoR_secext}: R-R intervals in seconds (with the starting zero).
   }
 }
 The function also writes an ascii text file "Rsec_data.txt" to the current work directory, which contains the \code{Rtrue_sec} variable and is required as input for the HRV analysis.
}
\description{
This function detects the R peaks from ECG data using wavelet decomposition.
The ECG data is provided in a data frame with two columns named \code{time} and \code{ecg.}
}
\examples{
load(system.file("extdata", "ecg.Rda", package="heartBeat"))
result <- heart_beat(ecg)
result <- heart_beat(ecg, SampleFreq = 250)
result <- heart_beat(ecg, SampleFreq = 360,  thr = 15)
}
\keyword{ECG,}
\keyword{beat,}
\keyword{detection}
\keyword{heart}

