% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat}
\alias{heartbeat}
\title{Conduct Heartbeat Analysis}
\usage{
heartbeat(
  data,
  type = c("counts", "standard", "votes"),
  threshold = 1,
  props = FALSE,
  controversial = TRUE
)
}
\arguments{
\item{data}{A data.frame.}

\item{type}{Type of output, default is "counts" which provides data.frame of counts, "standard" standardizes scores of original data.frame, and "votes" converts the standardization score into up-vote/down-vote/neutral.}

\item{threshold}{This denotes the range of "Neutral", or conversey, what deviation score above/below the base-rate is necessary to count as an up/down-vote. Default = 1.}

\item{props}{Default is false. If props = TRUE, will convert counts of each vote into proportions. Will only change output for type = "counts".}

\item{controversial}{Default is true. If controversial = FALSE, function will not provide additional controversy score for each item.}
}
\value{
Number of up-votes, down-votes, and neutral for each variable in data.frame \code{data}.
}
\description{
Conduct Heartbeat Analysis
}
\examples{
\dontrun{heartbeat(data, type = "counts", threshold = 0.5, props = TRUE, controversial = FALSE)}
}
