% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HRdistribution.R
\name{HRdistribution}
\alias{HRdistribution}
\title{Determine heart rate distribution}
\usage{
HRdistribution(Rall, signal, HRrest, age, training = FALSE)
}
\arguments{
\item{Rall}{data frame with information about (detected) heart beats with columns \itemize{\item \code{Rtrue_idx}: the indexes of R peaks in \code{signal}, \item \code{Rtrue_sec}: the time moments of \code{Rtrue_idx} in seconds, \item \code{RtoRext}: R-R intervals in number of samples (with the starting zero), \item \code{RtoR_secext}: R-R intervals in seconds (with the starting zero).},}

\item{signal}{vector of ECG signal data points,}

\item{HRrest}{numeric value of the monitored person's resting heart rate,}

\item{age}{numeric value of age of the monitored person's age}

\item{training}{a flag indicating, whether the monitored person was on trainning/exercise or not}
}
\value{
\code{beat_matrix}: data frame with columns coresponding to heart rate intervals.
 In each column the indexes of all heart beats that belong to the certain interval are writen.
 The number of columns is determined by the function and the number of rows is determined by the interval with the most heart beats.
 All other columns are appropriately filled with \code{NA} values.
}
\description{
This function determines instant heart rates, their distribution and a basic histogram based on the heart-beat detection output of the function \code{\link{heart_beat}}.
}
\examples{
load(system.file("extdata", "Rall.Rda", package="heartBeat"))
beat_matrix <- HRdistribution(Rall, data$ecg, 75, 30, training = FALSE)

load(system.file("extdata", "ecg.Rda", package="heartBeat"))
result <- heart_beat(ecg)
beat_matrix <- HRdistribution(result$Rall, data$ecg, 75, 30, training = FALSE)
}
\keyword{heart}
\keyword{histogram}
\keyword{instant}
\keyword{rate}
\keyword{rate,}

