% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_votes.R
\name{find_votes}
\alias{find_votes}
\title{Find up-votes and down-votes}
\usage{
find_votes(data, cols, threshold = 1)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble). See Methods, 
below, for more details.}
}
\description{
Find up-votes and down-votes
}
