% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat}
\alias{heartbeat}
\title{Conduct Heartbeat Analysis}
\usage{
heartbeat(
  data,
  type = c("counts", "standard", "votes"),
  threshold = 1,
  props = FALSE,
  controversial = TRUE
)
}
\arguments{
\item{data}{A numeric data frame or object that can be coerced into a data frame.}

\item{type}{Type of output. The default is "counts" which provides tallies the up/neutral/down votes of each item, "standard" standardizes each rating of the original data frame. "votes" converts the standardization score into up-vote/down-vote/neutral based on the threshold value.}

\item{threshold}{A number that denotes the range of "Neutral", or conversely, what deviation score above/below the base-rate is necessary to count as an up/down-vote. Default = 1.}

\item{props}{Default is false. If props = TRUE, it will convert counts of each vote into proportions. This Will only change output when type = "counts".}

\item{controversial}{Default is true. If controversial = FALSE, the function will not provide the controversy score for each item.}
}
\value{
Number of up-votes, down-votes, and neutral-votes for each variable in the data frame \code{data}.
}
\description{
Conduct Heartbeat Analysis
}
\examples{
\dontrun{heartbeat(data, type = "counts", threshold = 0.5, props = TRUE, controversial = FALSE)}
}
